requirements.txt
keras==2.13.1rc0
numpy==1.24.3
Pillow==9.4.0
Pillow==9.5.0
streamlit==1.22.0
